changed 'float' to 'int'

+++ 6/find_unique_numbers.py
@@ -3,7 +3,7 @@
 nums = []
 duplicates = {}
 for i in range(10):
-    nums.append(float(input(f"Enter number {i + 1}: ")))
+    nums.append(int(input(f"Enter number {i + 1}: ")))
     if nums[i] not in duplicates:
         duplicates[nums[i]] = 1
     else:
